requirements.txt
bitbucket-api
clime
PyGithub
simple-crypt
